Add authorization and conditional link rendering, add permissions to certain routes, update wishlist button functionality with animation and adding newly created lists

+++ thelastchapter/book_list.py
@@ -8,7 +8,9 @@ from thelastchapter.db import get_db
 
 from thelastchapter.account import get_books
 
-from thelastchapter.auth import login_required
+from thelastchapter.auth import (
+    login_required, check_permissions, permissions, check_list_ownership
+)
 
 bp = Blueprint('list', __name__, url_prefix='/lists')
 
@@ -20,10 +22,10 @@ def check_existence(list_id, book_id):
         abort(404)
     return None
 
-def res_format(dbStatus, message, list_id=None):
+def res_format(dbStatus, message, list_id=None, list_name=None):
     if not list_id:
         return { 'dbStatus': dbStatus, 'message': message }
-    return { 'dbStatus': dbStatus, 'message': message, 'list_id': list_id }
+    return { 'dbStatus': dbStatus, 'message': message, 'list_id': list_id, 'list_name': list_name }
 
 @bp.route('/create', methods=('POST',))
 @login_required
@@ -40,7 +42,7 @@ def create():
     cursor.execute('INSERT INTO book_lists (list_id, book_id) VALUES ( ?, ? )',
         (list_id, book_id))
     db.commit()
-    return res_format('success', 'New list successfully created!')
+    return res_format('success', 'New list successfully created!', list_id, name)
 
 @bp.route('/<int:list_id>')
 def display(list_id):
@@ -50,37 +52,39 @@ def display(list_id):
     if list_data is None:
         error = "Cannot find list."
     if error is None:
-        book_list = get_books(list_data)
-    if error is None and (book_list is None or book_list[2] is None):
+        book_data = get_books(list_data)
+    if error is None and (book_data is None or book_data[2] is None):
         error = "No books found"
     if error is not None:
         flash(error)
         return redirect(url_for('home'))
-    return render_template('list/display.html', book_list=book_list)
+    list_name, list_id, books, list_owner = book_data
+    print(books[0].keys())
+    return render_template('list/display.html', 
+        list_name=list_name, 
+        list_id=list_id,
+        books=books,
+        list_owner=list_owner
+    )
 
 @bp.route('/<int:list_id>/update', methods=('GET', 'POST'))
+@login_required
 def update(list_id):
-    db = get_db()
-    list_data = db.execute('SELECT * FROM list_names WHERE id = ?', (list_id,)).fetchone()
+    list_data = check_list_ownership(list_id)
     if list_data is None:
-        flash('List not found')
-        return redirect(url_for('account.display'))
+        return redirect(request.referrer)
     if request.method == 'POST':
+        db = get_db()
         name = request.form['name']
-        db.execute('UPDATE list_names SET name=? WHERE id = ?', (name, list_id))
+        db.execute('UPDATE list_names SET name=? WHERE id = ?', (name, list_id)).fetchone()
         db.commit()
         return redirect(url_for('list.display', list_id=list_id))
     return render_template('list/update.html', list_data=list_data)
 
 @bp.route('<int:list_id>/delete', methods=('POST',))
 def delete(list_id):
-    error = None
+    check_list_ownership(list_id)
     db = get_db()
-    list_data = db.execute('SELECT * FROM list_names WHERE id = ?', (list_id,)).fetchone()
-    if list_data is None:
-        error = "List not found!"
-        flash(error)
-        return redirect('home')
     db.execute('DELETE FROM book_lists WHERE list_id = ?', (list_id,))
     db.execute('DELETE FROM list_names WHERE id = ?', (list_id,))
     db.commit()
@@ -88,7 +92,7 @@ def delete(list_id):
 
 @bp.route('/<int:list_id>/<int:book_id>/remove', methods=("POST",))
 def remove(list_id, book_id):
-    check_existence(list_id, book_id)
+    check_list_ownership(list_id)
     db = get_db()
     db.execute('DELETE FROM book_lists WHERE list_id = ? AND book_id = ?', (list_id, book_id))
     db.commit()
@@ -96,10 +100,10 @@ def remove(list_id, book_id):
 
 @bp.route('/<int:list_id>/<int:book_id>/add', methods=("POST",))
 def add(list_id, book_id):
+    cont, data = check_list_ownership(list_id, True)
+    if not cont:
+        return data
     db = get_db()
-    book_list = db.execute('SELECT * FROM list_names WHERE id = ?', (list_id,)).fetchone()
-    if book_list is None:
-        return res_format('error', 'List not found')
     book = db.execute('SELECT * FROM books WHERE id = ?', (book_id,)).fetchone()
     if book is None:
         return res_format('error', 'Book not found')
